2021 updates

+++ src/main/java/com/cricket/wsipl2020/dto/MyCurrentRankResponse.java
@@ -7,6 +7,6 @@ import lombok.Setter;
 public class MyCurrentRankResponse {
 
     private String userName;
-    private Float points;
+    private Double points;
     private Integer rank;
 }
