+++ src/main/java/com/cricket/wsipl2020/model/Allocations.java
@@ -12,6 +12,6 @@ public class Allocations {
 
     @EmbeddedId
     AllocationsPK allocationsPK;
-    Float points;
+    Double points;
 
 }
