+++ src/main/java/com/cricket/wsipl2020/repository/ScheduleRepo.java
@@ -30,7 +30,7 @@ public interface ScheduleRepo extends CrudRepository<Schedule, Integer> {
 
     @Modifying @Transactional
     @Query(value = "Update schedule set winning_team = :winningTeam, max_points = :pointsEarned where game_num =:gameId",nativeQuery = true)
-    Integer updateWinningTeam(@Param("gameId") Integer gameId, @Param("winningTeam") String winningTeam,  @Param("pointsEarned") Float pointsEarned);
+    Integer updateWinningTeam(@Param("gameId") Integer gameId, @Param("winningTeam") String winningTeam,  @Param("pointsEarned") Double pointsEarned);
 
     @Query(value = "Select winning_team from schedule where game_num = :gameId",nativeQuery = true)
     String checkWinner(@Param("gameId") Integer gameId);
